+++ src/three/index.ts
@@ -16,32 +16,4 @@ THREE.BatchedMesh.prototype.computeBoundsTree = computeBatchedBoundsTree
 THREE.BatchedMesh.prototype.disposeBoundsTree = disposeBatchedBoundsTree
 THREE.BatchedMesh.prototype.raycast = acceleratedRaycast
 
-/**
- * Recursively enables BVH for all meshes in the object
- */
-export function enableBVH(object: THREE.Object3D) {
-	object.traverse((child) => {
-		if (child instanceof THREE.Mesh && child.geometry) {
-			child.geometry.computeBoundsTree()
-		}
-		if (child instanceof THREE.BatchedMesh) {
-			child.computeBoundsTree()
-		}
-	})
-}
-
-/**
- * Recursively disposes BVH for all meshes in the object
- */
-export function disposeBVH(object: THREE.Object3D) {
-	object.traverse((child) => {
-		if (child instanceof THREE.Mesh && child.geometry) {
-			child.geometry.disposeBoundsTree()
-		}
-		if (child instanceof THREE.BatchedMesh) {
-			child.disposeBoundsTree()
-		}
-	})
-}
-
 export default THREE