+++ src/three/three.d.ts
@@ -5,14 +5,22 @@ import {
 	disposeBatchedBoundsTree,
 	acceleratedRaycast
 } from 'three-mesh-bvh'
-import type THREE from '.'
 
-interface Object3DUserData {
+interface MxObjectUserData {
 	isShadable?: boolean
 	isHelper?: boolean
 	isSceneLight?: boolean
 	skipRaycast?: boolean
 	isSelectable?: boolean
+	mixeur?: {
+		isShadable?: boolean
+		isHelper?: boolean
+		isSceneLight?: boolean
+		skipRaycast?: boolean
+		isSelectable?: boolean
+		visibleIn?: ('wireframe' | 'solid' | 'preview' | 'rendered' | 'export')[]
+		hideInOutliner?: boolean
+	}
 	[key: string]: unknown
 }
 
@@ -25,22 +33,11 @@ declare module 'three' {
 
 	interface Mesh {
 		raycast: typeof acceleratedRaycast
-		userData: {
-			originalMaterial?: THREE.Material | THREE.Material[]
-		} & Object3DUserData
 	}
 
 	interface BatchedMesh {
 		computeBoundsTree: typeof computeBatchedBoundsTree
 		disposeBoundsTree: typeof disposeBatchedBoundsTree
 		raycast: typeof acceleratedRaycast
-	}
-
-	interface Object3D {
-		userData: Object3DUserData
-	}
-
-	interface Light {
-		userData: Object3DUserData
 	}
 }